@@ -12,56 +12,56 @@
     "postbuild": "node -e \"fs.copyFileSync('dist/index.html','dist/404.html')\""
   },
   "dependencies": {
-    "@radix-ui/react-dialog": "^1.0.5",
-    "@radix-ui/react-dropdown-menu": "^2.0.6",
-    "@radix-ui/react-slot": "^1.0.2",
-    "@servicestack/client": "^2.1.2",
-    "class-variance-authority": "^0.7.0",
+    "@radix-ui/react-dialog": "^1.1.13",
+    "@radix-ui/react-dropdown-menu": "^2.1.14",
+    "@radix-ui/react-slot": "^1.2.2",
+    "@servicestack/client": "^2.1.11",
+    "class-variance-authority": "^0.7.1",
     "classnames": "^2.5.1",
-    "clsx": "^2.1.0",
-    "lucide-react": "^0.330.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-helmet-async": "^2.0.4",
-    "react-router": "^6.22.0",
-    "react-router-dom": "^6.22.0",
-    "swr": "^2.2.4",
-    "tailwind-merge": "^2.2.1",
+    "clsx": "^2.1.1",
+    "lucide-react": "^0.510.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-router": "^7.6.0",
+    "react-router-dom": "^7.6.0",
+    "swr": "^2.3.3",
+    "tailwind-merge": "^3.3.0",
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
-    "@iconify/react": "^4.1.1",
-    "@mdx-js/rollup": "^3.0.1",
-    "@tailwindcss/forms": "^0.5.7",
-    "@tailwindcss/typography": "^0.5.10",
-    "@types/mdx": "^2.0.11",
-    "@types/node": "^20.11.20",
-    "@types/react": "^18.2.56",
-    "@types/react-dom": "^18.2.19",
-    "@types/react-helmet-async": "^1.0.3",
+    "@iconify/react": "^6.0.0",
+    "@mdx-js/rollup": "^3.1.0",
+    "@tailwindcss/forms": "^0.5.10",
+    "@tailwindcss/postcss": "^4.1.6",
+    "@tailwindcss/typography": "^0.5.16",
+    "@tailwindcss/vite": "^4.1.6",
+    "@types/mdx": "^2.0.13",
+    "@types/node": "^22.15.17",
+    "@types/react": "^19.1.4",
+    "@types/react-dom": "^19.1.5",
     "@types/react-router-dom": "^5.3.3",
     "@types/remark-prism": "^1.3.7",
-    "@typescript-eslint/eslint-plugin": "^6.21.0",
-    "@typescript-eslint/parser": "^6.21.0",
-    "@vitejs/plugin-react": "^4.2.1",
-    "autoprefixer": "^10.4.17",
-    "eslint": "^8.56.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.5",
-    "hastscript": "^9.0.0",
-    "postcss": "^8.4.35",
+    "@typescript-eslint/eslint-plugin": "^8.32.1",
+    "@typescript-eslint/parser": "^8.32.1",
+    "@vitejs/plugin-react": "^4.4.1",
+    "autoprefixer": "^10.4.21",
+    "eslint": "^9.26.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.20",
+    "hastscript": "^9.0.1",
+    "postcss": "^8.5.3",
     "rehype-raw": "^7.0.0",
-    "rehype-stringify": "^10.0.0",
-    "remark-directive": "^3.0.0",
+    "rehype-stringify": "^10.0.1",
+    "remark-directive": "^4.0.0",
     "remark-frontmatter": "^5.0.0",
-    "remark-gfm": "^4.0.0",
+    "remark-gfm": "^4.0.1",
     "remark-parse": "^11.0.0",
     "remark-prism": "^1.3.6",
-    "tailwindcss": "^3.4.1",
-    "typescript": "^5.2.2",
-    "vite": "^5.1.0",
-    "vite-plugin-pages": "^0.32.0",
+    "tailwindcss": "^4.1.6",
+    "typescript": "^5.8.3",
+    "vite": "^6.3.5",
+    "vite-plugin-pages": "^0.33.0",
     "vite-plugin-press": "^1.0.10",
-    "vite-plugin-svgr": "^4.2.0"
+    "vite-plugin-svgr": "^4.3.0"
   }
 }
